requirements.txt
pydantic>=2.11.7
requests>=2.32.4
rfc3986>=2.0.0
mcp>=1.13.0
pyyaml>=6.0